+++ lib/OrdinaryDiffEqStabilizedIRK/Project.toml
@@ -1,5 +1,5 @@
-name = "OrdinaryDiffEqStabilizedRK"
-uuid = "358294b1-0aab-51c3-aafe-ad5ab194a2ad"
+name = "OrdinaryDiffEqStabilizedIRK"
+uuid = "e3e12d00-db14-5390-b879-ac3dd2ef6296"
 authors = ["ParamThakkar123 <paramthakkar864@gmail.com>"]
 version = "1.0.0"
