+++ lib/OrdinaryDiffEqStabilizedIRK/src/algorithms.jl
@@ -1,0 +1,22 @@
+struct IRKC{CS, AD, F, F2, P, FDT, ST, CJ, K, T, E} <:
+       OrdinaryDiffEqNewtonAdaptiveAlgorithm{CS, AD, FDT, ST, CJ}
+    linsolve::F
+    nlsolve::F2
+    precs::P
+    κ::K
+    tol::T
+    extrapolant::Symbol
+    controller::Symbol
+    eigen_est::E
+end
+
+function IRKC(; chunk_size = Val{0}(), autodiff = Val{true}(), standardtag = Val{true}(),
+        concrete_jac = nothing, diff_type = Val{:forward},
+        linsolve = nothing, precs = DEFAULT_PRECS, nlsolve = NLNewton(), κ = nothing,
+        tol = nothing,
+        extrapolant = :linear, controller = :Standard, eigen_est = nothing)
+    IRKC{_unwrap_val(chunk_size), _unwrap_val(autodiff), typeof(linsolve), typeof(nlsolve),
+        typeof(precs), diff_type, _unwrap_val(standardtag), _unwrap_val(concrete_jac),
+        typeof(κ), typeof(tol), typeof(eigen_est)}(linsolve, nlsolve, precs, κ, tol,
+        extrapolant, controller, eigen_est)
+end
