+++ lib/OrdinaryDiffEqStabilizedIRK/test/runtests.jl
@@ -1,0 +1,3 @@
+using SafeTestsets
+
+@time @safetestset "IRKC Tests" include("irkc_tests.jl")
